requirements.txt
fastapi==0.115.12
numpy==2.2.3
oss2==2.19.1
pydantic==2.11.4
python-dotenv==1.1.0
SQLAlchemy==2.0.40
starlette==0.46.2
torch==2.6.0
uvicorn==0.34.2
asyncpg==0.30.0
